Add Attach File to FrontEnd Dynamic and Stand Form

+++ app/core/services/operations/Requests/RequestService.php
@@ -28,7 +28,9 @@ class RequestService
         /** @var Request $request */
         $request = $this->requests->get($id);
         $form = $request->requestForm;
-        $form->delete();
+        if ($form) {
+            $form->delete();
+        }
         $this->requests->remove($request);
     }
 }
